// packages/api-client/src/types/responseErrorResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Todo API
 * A simple todo API built with Go and Gin framework
 * OpenAPI spec version: 1.0
 */
import type { ResponseValidationErrorDetail } from './responseValidationErrorDetail'

export interface ResponseErrorResponse {
  /** バリデーションエラーがあるときはこちらでまとめて返す */
  details: ResponseValidationErrorDetail[]
  message: string
  status: number
}